requirements.txt
asgiref==3.2.10
autopep8==1.5.3
Django==3.0.8
factory-boy==2.12.0
Faker==4.1.1
pycodestyle==2.6.0
python-dateutil==2.8.1
pytz==2020.1
PyYAML==5.3.1
six==1.15.0
sqlparse==0.3.1
text-unidecode==1.3
toml==0.10.1
